Make a_filename const char* in Load/Save methods
This affects the Genome, NeuralNetwork, Parameters and Population,
classes.  Since a_filename is not modified by the Load/Save methods, it
makes sense for it to be const. Having a_filename be const simplifies
interaction with callers that use std::strings.

The changes are backwards compatible - you can still pass a char* for
a_filename.